fix(simulation): cast gz airspeed temperature to double
The macOS build has been failing since #28842: clang reports the
implicit float to double conversion under -Wdouble-promotion, which the
build treats as an error. GCC does not warn for argument conversions,
so the Linux builds passed.

+++ src/modules/simulation/gz_plugins/airspeed/AirSpeed.cpp
@@ -146,7 +146,7 @@ void AirSpeed::PreUpdate(const gz::sim::UpdateInfo &_info,
 			       (double)diff_pressure_noise;
 	gz::msgs::AirSpeed airspeed_msg;
 	airspeed_msg.set_diff_pressure(diff_pressure * 100.0);
-	airspeed_msg.set_temperature(temperature_local); // [K]
+	airspeed_msg.set_temperature((double)temperature_local); // [K]
 	_pub.Publish(airspeed_msg);
 }
